add cluster_id_mgr to handle cluster_id
