+++ be/src/storage/CMakeLists.txt
@@ -156,4 +156,5 @@ add_library(Storage STATIC
     vertical_compaction_task.cpp
     compaction_task_factory.cpp
     base_and_cumulative_compaction_policy.cpp
+    cluster_id_mgr.cpp
 )
