+++ be/src/storage/data_dir.cpp
@@ -57,16 +57,16 @@ namespace starrocks {
 static const char* const kMtabPath = "/etc/mtab";
 static const char* const kTestFilePath = "/.testfile";
 
-DataDir::DataDir(std::string path, TStorageMedium::type storage_medium, TabletManager* tablet_manager,
+DataDir::DataDir(const std::string& path, TStorageMedium::type storage_medium, TabletManager* tablet_manager,
                  TxnManager* txn_manager)
-        : _path(std::move(path)),
+        : _path(path),
           _available_bytes(0),
           _disk_capacity_bytes(0),
           _storage_medium(storage_medium),
           _is_used(false),
           _tablet_manager(tablet_manager),
           _txn_manager(txn_manager),
-          _cluster_id(-1),
+          _cluster_id_mgr(std::make_shared<ClusterIdMgr>(path)),
           _current_shard(0) {}
 
 DataDir::~DataDir() {
@@ -84,7 +84,7 @@ Status DataDir::init(bool read_only) {
     }
 
     RETURN_IF_ERROR_WITH_WARN(update_capacity(), "update_capacity failed");
-    RETURN_IF_ERROR_WITH_WARN(_init_cluster_id(), "_init_cluster_id failed");
+    RETURN_IF_ERROR_WITH_WARN(_cluster_id_mgr->init(), "_cluster_id_mgr init failed");
     RETURN_IF_ERROR_WITH_WARN(_init_data_dir(), "_init_data_dir failed");
     RETURN_IF_ERROR_WITH_WARN(_init_tmp_dir(), "_init_tmp_dir failed");
     RETURN_IF_ERROR_WITH_WARN(_init_meta(read_only), "_init_meta failed");
@@ -96,79 +96,6 @@ Status DataDir::init(bool read_only) {
 void DataDir::stop_bg_worker() {
     _stop_bg_worker = true;
     _cv.notify_one();
-}
-
-Status DataDir::_init_cluster_id() {
-    std::string cluster_id_path = _path + CLUSTER_ID_PREFIX;
-    if (access(cluster_id_path.c_str(), F_OK) != 0) {
-        int fd = open(cluster_id_path.c_str(), O_RDWR | O_CREAT, S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP);
-        if (fd < 0 || close(fd) < 0) {
-            RETURN_IF_ERROR_WITH_WARN(Status::IOError(strings::Substitute("failed to create cluster id file $0, err=$1",
-                                                                          cluster_id_path, errno_to_string(errno))),
-                                      "create file failed");
-        }
-    }
-
-    // obtain lock of all cluster id paths
-    FILE* fp = nullptr;
-    fp = fopen(cluster_id_path.c_str(), "r+b");
-    if (fp == nullptr) {
-        RETURN_IF_ERROR_WITH_WARN(
-                Status::IOError(strings::Substitute("failed to open cluster id file $0", cluster_id_path)),
-                "open file failed");
-    }
-
-    DeferOp close_fp([&]() {
-        fclose(fp);
-        fp = nullptr;
-    });
-
-    int lock_res = flock(fp->_fileno, LOCK_EX | LOCK_NB);
-    if (lock_res < 0) {
-        RETURN_IF_ERROR_WITH_WARN(
-                Status::IOError(strings::Substitute("failed to flock cluster id file $0", cluster_id_path)),
-                "flock file failed");
-    }
-
-    // obtain cluster id of all root paths
-    auto st = _read_cluster_id(cluster_id_path, &_cluster_id);
-    return st;
-}
-
-Status DataDir::_read_cluster_id(const std::string& path, int32_t* cluster_id) {
-    RETURN_IF_ERROR(_add_version_info_to_cluster_id(path));
-
-    std::fstream fs(path.c_str(), std::fstream::in);
-    if (!fs.is_open()) {
-        RETURN_IF_ERROR_WITH_WARN(Status::IOError(strings::Substitute("failed to open cluster id file $0", path)),
-                                  "open file failed");
-    }
-
-    std::string cluster_id_str;
-    fs >> cluster_id_str;
-    fs.close();
-    int32_t tmp_cluster_id = -1;
-    if (!cluster_id_str.empty()) {
-        size_t pos = cluster_id_str.find('-');
-        if (pos != std::string::npos) {
-            tmp_cluster_id = std::stoi(cluster_id_str.substr(0, pos));
-        } else {
-            tmp_cluster_id = std::stoi(cluster_id_str);
-        }
-    }
-
-    if (tmp_cluster_id == -1 && (fs.rdstate() & std::fstream::eofbit) != 0) {
-        *cluster_id = -1;
-    } else if (tmp_cluster_id >= 0 && (fs.rdstate() & std::fstream::eofbit) != 0) {
-        *cluster_id = tmp_cluster_id;
-    } else {
-        RETURN_IF_ERROR_WITH_WARN(Status::Corruption(strings::Substitute(
-                                          "cluster id file $0 is corrupt. [id=$1 eofbit=$2 failbit=$3 badbit=$4]", path,
-                                          tmp_cluster_id, fs.rdstate() & std::fstream::eofbit,
-                                          fs.rdstate() & std::fstream::failbit, fs.rdstate() & std::fstream::badbit)),
-                                  "file content is error");
-    }
-    return Status::OK();
 }
 
 Status DataDir::_init_data_dir() {
@@ -202,57 +129,7 @@ Status DataDir::_init_meta(bool read_only) {
 }
 
 Status DataDir::set_cluster_id(int32_t cluster_id) {
-    if (_cluster_id != -1) {
-        if (_cluster_id == cluster_id) {
-            return Status::OK();
-        }
-        LOG(ERROR) << "going to set cluster id to already assigned store, cluster_id=" << _cluster_id
-                   << ", new_cluster_id=" << cluster_id;
-        return Status::InternalError("going to set cluster id to already assigned store");
-    }
-    return _write_cluster_id_to_path(_cluster_id_path(), cluster_id);
-}
-
-Status DataDir::_write_cluster_id_to_path(const std::string& path, int32_t cluster_id) {
-    std::fstream fs(path.c_str(), std::fstream::out);
-    if (!fs.is_open()) {
-        LOG(WARNING) << "fail to open cluster id path. path=" << path;
-        return Status::InternalError("IO Error");
-    }
-    fs << cluster_id;
-    fs << "-" << std::string(STARROCKS_VERSION);
-    fs.close();
-    return Status::OK();
-}
-
-// This function is to add version info into file named cluster_id
-// This feacture is used to restrict the degrading from StarRocks-1.17.2 to lower version
-// Because the StarRocks with lower version cannot read the file written by RocksDB-6.22.1
-// This feature takes into effect after StarRocks-1.17.2
-// Without this feature, staring BE will be failed
-Status DataDir::_add_version_info_to_cluster_id(const std::string& path) {
-    std::fstream in_fs(path.c_str(), std::fstream::in);
-    if (!in_fs.is_open()) {
-        RETURN_IF_ERROR_WITH_WARN(Status::IOError(strings::Substitute("failed to open cluster id file $0", path)),
-                                  "open file failed");
-    }
-    std::string cluster_id_str;
-    in_fs >> cluster_id_str;
-    in_fs.close();
-
-    if (cluster_id_str.empty() || cluster_id_str.find('-') != std::string::npos) {
-        return Status::OK();
-    }
-
-    std::fstream out_fs(path.c_str(), std::fstream::out);
-    if (!out_fs.is_open()) {
-        RETURN_IF_ERROR_WITH_WARN(Status::IOError(strings::Substitute("failed to open cluster id file $0", path)),
-                                  "open file failed");
-    }
-    out_fs << cluster_id_str;
-    out_fs << "-" << std::string(STARROCKS_VERSION);
-    out_fs.close();
-    return Status::OK();
+    return _cluster_id_mgr->set_cluster_id(cluster_id);
 }
 
 void DataDir::health_check() {
